vhost fuzzing support for http

+++ src/filters/content_length.rs
@@ -2,8 +2,8 @@ use std::fmt::{Display, Formatter};
 
 #[derive(Clone, Debug, PartialEq)]
 pub enum FilterContentLength {
-    Separate(Vec<u32>),
-    Range(u32, u32),
+    Separate(Vec<usize>),
+    Range(usize, usize),
     Empty,
 }
 
@@ -17,20 +17,20 @@ impl From<&str> for FilterContentLength {
     fn from(value: &str) -> Self {
         if value.contains("-") {
             return Self::from_nums(value.split("-")
-                .map(|v| v.parse::<u32>())
+                .map(|v| v.parse::<usize>())
                 .flatten()
                 .collect());
         }
 
         FilterContentLength::Separate(value.split(",")
-            .map(|v| v.parse::<u32>())
+            .map(|v| v.parse::<usize>())
             .flatten()
             .collect())
     }
 }
 
 impl FilterContentLength {
-    pub fn matches(&self, length: u32) -> bool {
+    pub fn matches(&self, length: usize) -> bool {
         match self {
             FilterContentLength::Empty => false,
             FilterContentLength::Separate(v) => v.contains(&length),
@@ -38,8 +38,8 @@ impl FilterContentLength {
         }
     }
 
-    fn from_nums(nums: Vec<u32>) -> FilterContentLength {
-        let nums: [u32; 2] = nums.try_into()
+    fn from_nums(nums: Vec<usize>) -> FilterContentLength {
+        let nums: [usize; 2] = nums.try_into()
             .unwrap_or_else(|_| panic!("expected 2 values in content length range"));
 
         if nums[0] < nums[1] {
